requirements.txt
flask
flask-cors
python-dotenv
playwright
pandas
tldextract
langdetect
openai
